remove minor bugs and authentication

+++ src/app/(protected)/admin/(components)/EditItemComp.tsx
@@ -1,14 +1,19 @@
 import ImageComp from "@/app/(components)/ImageComp";
-import { Item } from "@/app/types/types";
+import { Item, UserRole } from "@/app/types/types";
 import { Box, Button, Card, Typography } from "@mui/material";
+import { useSession } from "next-auth/react";
 import React from "react";
 
 const EditItemComp = ({ item }: { item: Item }) => {
+  const { data: session } = useSession();
   const editItem = (id: number) => {
     window.location.href = `/admin/handleItem/${id}`;
   };
 
   const removeItem = async (id: number) => {
+    if (!session || session.user.role !== UserRole.ADMIN)
+      throw new Error("you are not authorized");
+
     const res = await fetch(`http://localhost:5000/api/products/${id}`, {
       method: "DELETE",
       headers: {
@@ -47,7 +52,7 @@ const EditItemComp = ({ item }: { item: Item }) => {
         >
           <Button onClick={() => editItem(item.id)}>edit</Button>
           <Button onClick={() => removeItem(item.id)}>remove</Button>
-          <Typography variant={"h6"}>${item.price.toFixed(2)}</Typography>
+          <Typography variant={"h6"}>${(+item.price).toFixed(2)}</Typography>
         </Card>
       </Card>
     </Box>
